refactor(tic-tac-toe): Move components to single files
