+++ react-tic-tac-toe/src/components/Square/index.js
@@ -1,0 +1,7 @@
+export const Square = (props) => (
+  <button className="square" onClick={props.onClick}>
+    {props.value}
+  </button>
+);
+
+export default Square;
