+++ react-tic-tac-toe/src/index.js
@@ -2,137 +2,17 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 
-const Square = (props) => (
-  <button className="square" onClick={props.onClick}>
-    {props.value}
-  </button>
-);
+import Game from './components/Game';
 
-class Board extends React.Component {
-  renderSquare(i) {
-    return (
-      <Square
-        value={this.props.squares[i]}
-        onClick={() => this.props.onClick(i)}
-      />
-    );
-  }
+/* -- TODO --
 
-  render() {
-    return (
-      <div>
-        <div className="board-row">
-          {this.renderSquare(0)}
-          {this.renderSquare(1)}
-          {this.renderSquare(2)}
-        </div>
-        <div className="board-row">
-          {this.renderSquare(3)}
-          {this.renderSquare(4)}
-          {this.renderSquare(5)}
-        </div>
-        <div className="board-row">
-          {this.renderSquare(6)}
-          {this.renderSquare(7)}
-          {this.renderSquare(8)}
-        </div>
-      </div>
-    );
-  }
-}
-
-class Game extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      history: [{ squares: Array(9).fill(null) }],
-      xIsNext: true,
-      stepNumber: 0,
-    };
-  }
-
-  jumpTo(step) {
-    this.setState({
-      stepNumber: step,
-      xIsNext: (step % 2) === 0,
-    });
-  }
-
-  handleClick(i) {
-    const { history, xIsNext, stepNumber } = this.state;
-    const { squares } = history[stepNumber];
-
-    if (calculateWinner(squares) || squares[i]) {
-      return;
-    }
-
-    const newSquares = squares.slice();
-
-    newSquares[i] = xIsNext ? 'X' : 'O';
-
-    this.setState({
-      history: [...history.slice(0, stepNumber + 1), { squares: newSquares }],
-      stepNumber: stepNumber + 1,
-      xIsNext: !xIsNext,
-    });
-  }
-
-  render() {
-    const { history, xIsNext, stepNumber } = this.state;
-    const { squares } = history[stepNumber];
-
-    const winner = calculateWinner(squares);
-    const nextPlayer = xIsNext ? 'X' : 'O';
-    const status = winner ? `Winner: ${winner}` : `Next player: ${nextPlayer}`;
-
-    const moves = history.map((step, move) => {
-      const desc = move ? `Go to move ${move}` : `Go to game start`;
-
-      return (
-        <li key={move}>
-          <button onClick={() => this.jumpTo(move)}>{desc}</button>
-        </li>
-      );
-    });
-
-    return (
-      <div className="game">
-        <div className="game-board">
-          <Board squares={squares} onClick={(i) => this.handleClick(i)} />
-        </div>
-        <div className="game-info">
-          <div>{ status }</div>
-          <ol>{ moves }</ol>
-        </div>
-      </div>
-    );
-  }
-}
-
-function calculateWinner(squares) {
-  const lines = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6],
-  ];
-
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      return squares[a];
-    }
-  }
-
-  return null;
-}
-
-// ========================================
+  - Display the location for each move in the format (col, row) in the move history list.
+  - Bold the currently selected item in the move list.
+  - Rewrite Board to use two loops to make the squares instead of hardcoding them.
+  - Add a toggle button that lets you sort the moves in either ascending or descending order.
+  - When someone wins, highlight the three squares that caused the win.
+  - When no one wins, display a message about the result being a draw.
+*/
 
 ReactDOM.render(
   <Game />,
